requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
cloudscraper==1.2.71
idna==3.10
pixivpy3==3.7.5
pyparsing==3.2.1
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
typing_extensions==4.12.2
urllib3==2.3.0
